feat: regenerate protobuf, add E2E integration tests, fix autocli
- Regenerate liquidityvault .pb.go files with proper file descriptors
  (replaces hand-written stubs that had nil Descriptor() methods,
  which caused chain startup panic in RegisterMsgServiceDesc)
- Add query.pb.gw.go gateway stub for RegisterQueryHandlerClient
- Add PositionalArgs to autocli tx commands (register-validator,
  deposit-to-vault, set-delegator-reward-percent) and query commands
- Fix genesis.go and genesis_test.go for int64 field types (matching
  regenerated proto which uses int64 for check height fields)
- Add comprehensive E2E integration tests (integration_test.go):
  - QueryPositionValue with successful mock wasm responses
  - Full lifecycle: register → deposit → query value → score → ranking
  - Deposit and total vault value calculation
  - Multiple validators competing in same pool
  - BeginBlock-driven full cycle
  - Zero total liquidity edge case
- Verified chain binary builds, starts, and processes transactions:
  - register-validator (LIQUIDITY type) succeeded on live local testnet
  - set-delegator-reward-percent succeeded on live local testnet
  - Vault data confirmed present via raw ABCI query

+++ x/liquidityvault/module/genesis.go
@@ -30,10 +30,10 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 
 	// Restore check heights
 	if genState.LastSimpleCheckHeight > 0 {
-		k.SetLastSimpleCheckHeight(ctx, int64(genState.LastSimpleCheckHeight))
+		k.SetLastSimpleCheckHeight(ctx, genState.LastSimpleCheckHeight)
 	}
 	if genState.LastComplexCheckHeight > 0 {
-		k.SetLastComplexCheckHeight(ctx, int64(genState.LastComplexCheckHeight))
+		k.SetLastComplexCheckHeight(ctx, genState.LastComplexCheckHeight)
 	}
 }
 
@@ -41,8 +41,8 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 	genesis := types.DefaultGenesis()
 	genesis.Params = k.GetParams(ctx)
-	genesis.LastSimpleCheckHeight = uint64(k.GetLastSimpleCheckHeight(ctx))
-	genesis.LastComplexCheckHeight = uint64(k.GetLastComplexCheckHeight(ctx))
+	genesis.LastSimpleCheckHeight = k.GetLastSimpleCheckHeight(ctx)
+	genesis.LastComplexCheckHeight = k.GetLastComplexCheckHeight(ctx)
 
 	// Export all validator records
 	vaults := k.GetAllVaults(ctx)
